폰트수정/홈->상세보기 이동
