+++ lib/common/storage.dart
@@ -1,8 +1,8 @@
 String? firebasetoken='임시';
 String? kakaotoken='임시';
-String? useremail='4@gmail.com';
-String latitude='';
-String longtitude='';
+String? useremail='7@gmail.com';
+String latitude='12.394';
+String longtitude='394.393';
 
 //서버 url
 String kakaonativekey = 'd54cd76337470f87b093bfdadfa53292'; //카카오토큰
