+++ lib/common/uri.dart
@@ -1,2 +1,3 @@
-final posturl=Uri.parse('http://116.44.231.155:8080/api/post');
-final url='http://116.44.231.155:8080';
+final posturl=Uri.parse('http://124.60.219.83:8080/api/post');
+final url='http://124.60.219.83:8080';
+//124.60.219.83
