+++ lib/common/util.dart
@@ -1,5 +1,10 @@
+import 'dart:ui';
+
 import 'package:dorandoran/common/storage.dart';
+import 'package:flutter/src/painting/text_style.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:geolocator/geolocator.dart';
+import 'package:google_fonts/google_fonts.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 //00 시간 출력설정
@@ -77,3 +82,11 @@ void getlocation() async {
     longtitude = position.longitude.toString();
 }
 
+//스타일가져오기
+TextStyle selectfont(String font, String fontColor, int fontSize, int fontBold){
+  Color color=fontColor=="black" ? Color(0xFF000000):Color(0xFFFFFFFF);
+  TextStyle style = GoogleFonts.getFont(font, textStyle: TextStyle(fontSize: fontSize.sp, color:color, fontWeight: FontWeight.bold));
+  print("반환");
+  return style;
+
+}
