+++ lib/texting/get/model/postcard_detaril.dart
@@ -1,4 +1,3 @@
-
 class postcardDetail {
   final String content;
   final String postTime;
@@ -10,6 +9,10 @@ class postcardDetail {
   final String? postNickname;
   final List<dynamic>? commentDetailDto;
   final List<dynamic>? postHashes;
+  final String font;
+  final String fontColor;
+  final int fontSize;
+  final int fontBold;
 
   postcardDetail({
     required this.content,
@@ -22,20 +25,27 @@ class postcardDetail {
     required this.postNickname,
     required this.commentDetailDto,
     required this.postHashes,
+    required this.font,
+    required this.fontColor,
+    required this.fontSize,
+    required this.fontBold,
   });
 
   factory postcardDetail.fromJson(Map<String, dynamic> json) {
     return postcardDetail(
-      content: json["content"],
-      postTime: json["postTime"],
-      location: json["location"],
-      postLikeCnt: json["postLikeCnt"],
-      postLikeResult: json["likeResult"] ?? false,
-      commentCnt: json["commentCnt"],
-      backgroundPicUri: json["backgroundPicUri"],
-      postNickname: json["postNickname"],
-      commentDetailDto: json["commentDetailDto"],
-      postHashes: json["postHashes"],
-    );
+        content: json["content"],
+        postTime: json["postTime"],
+        location: json["location"],
+        postLikeCnt: json["postLikeCnt"],
+        postLikeResult: json["likeResult"] ?? false,
+        commentCnt: json["commentCnt"],
+        backgroundPicUri: json["backgroundPicUri"],
+        postNickname: json["postNickname"],
+        commentDetailDto: json["commentDetailDto"],
+        postHashes: json["postHashes"],
+        font: json["font"],
+        fontColor: json["fontColor"],
+        fontSize: json["fontSize"],
+        fontBold: json["fontBold"]);
   }
 }
